Fixed a few test app type conversions that Visual Studio complains about.

+++ test/testdyngl.c
@@ -138,9 +138,9 @@ int main(int argc,char *argv[])
 
 	for(i=0;i<NB_PIXELS;i++)
 	{
-		pixels[3*i]=rand()%250-125;
-		pixels[3*i+1]=rand()%250-125;
-		pixels[3*i+2]=rand()%250-125;
+		pixels[3*i]=(GLfloat) (rand()%250-125);
+		pixels[3*i+1]=(GLfloat) (rand()%250-125);
+		pixels[3*i+2]=(GLfloat) (rand()%250-125);
 	}
 	
 	f.glViewport(0,0,640,480);
@@ -166,7 +166,7 @@ int main(int argc,char *argv[])
 	f.glEnable(GL_FOG);
 	f.glFogf(GL_FOG_START,-500);
 	f.glFogf(GL_FOG_END,500);
-	f.glFogf(GL_FOG_DENSITY,0.005);
+	f.glFogf(GL_FOG_DENSITY,0.005f);
 	
 	do
 	{
